Bugfix: Keep track of ensured/missing equalities in tree interfaces.
Previously, we only kept track of disequalities.

This led to wrong results in case the the matching of the LHS to the
RHS is only allowed if certain equalities hold for the LHS, but these
equalities were not present.
For example, when matching

  x1 -> (x2, x3) |= x1 -> (null,null)

it is necessary to record that x2 = null and x3 = null are missing
constraints that need to be provided in the context to make the
entailment hold.

See examples in examples/entailment/equalities for various benchmarks
that exhibit this behavior.