Immediately discard extension types that don't have enough names.
For a tree interface to be composable, we need nonplaceholder var
labels for the following parameters of all nodes:
- the root parameter of every node (if a root parameter is specified)
- all parameters that are marked as used

These properties are now checked in the entailment automaton and
ETypes that violate any of these properties are immediately discraded.

In a rooted SID, the second condition actually subsumes the first. In
the not-necessarily-rooted part of the SID that corresponds to the
top-level query, we can have root parameters that are not marked as
used, e.g. if we compute the usage info for

  x1 -> x2 * ls(y1, x3),

y1 is an unused root parameter.