Bugfix: Don't compute targets for transitions with inconsistent source instantiations.
For example, if the label of a transition contains a call P(y,null),
states in which the second free variable occurs in the root position
of a node of a tree interface (and hence null must be allocated) must
not be considered as source states, because there are no completions
of that tree interface to a full tree in which null can actually be
allocated.

See benchmarks for an example that now works after this change.