Fix: Consider all non-progress rules that can be applied.
This makes a difference if the top-level predicate of the progress
normal form itself contains a copy of a non-progress rule of the
underlying SID. This is the case, for example, if the RHS is a single
call to a non-progress SID but with bound variables:

  ∃α1 . sll(x1,α1)

w.r.t. SID

  sll(x1, x2) <= x1 ↦ x2
  sll(x1, x2) <= ∃y . x1 ↦ y * sll(y,x2)

is rewritten to

  rhs1(x1) <= ∃_1 . x1 ↦ _1
  rhs1(x1) <= ∃_1 ∃_2 . x1 ↦ _2 * sll(_2,_1) ;
  sll(x1, x2) <= x1 ↦ x2
  sll(x1, x2) <= ∃y . x1 ↦ y * sll(y,x2)

In this case, both the second rule of rhs1 and of sll have to be
considered when no-progress rules are applied.