Bugfix: Consider all options to merge zero or more tree interfaces.
Previously, we always performed a "maximal merge", merging as many
tree interfaces as possible when composing two extension
types. However, such aggressive merging may lose information about cut
points: We lose the information which free variables are labeling the
nodes at which we merge, which may be necessary to show the
entailment.

See examples/entailment/various/list-segments-different-order.hrs for
a benchmark that relies on this new definition of composition.