Fixed problem with diverging subprocesses in the new "integrated" benchmarking approach.
To produce a table with reliable timeout and error information, we
call each tool once on the file and only if that succeeds call JMH.

However, the subprocesses in the initial calls were not terminated
properly on timeouts, leading to an accumulating number
of (busyworking?) SLIDE processes. This is fixed now through an
explicit SIGKILL (destroy call) after the timeout.