Annotation with root parameters is now a fixed-point computation.
In SIDs that violate progress, we can now also discover the root
parameters of predicates that have some rules that have calls but
don't allocate memory, by propagating the root parameters of the
called predicates (discovered in previous annotation iterations).

This works for union types (which occur in the SL-COMP benchmarks) as
well as other well-founded dependency hierarchies between predicates.