The forget operation for profiles now performs a full filter.
Instead of just dropping decompositions if we don't have names for
root parameters, we now drop them if we don't have name for any
arbitrary used parameter. (This corresponds to the TACAS forget
operation generalized to the fragment with pure formulas.)

This drastically reduces the number of decompositions in some
complicated SIDs, e.g. grids.