Moving usage info and pure constraints from contexts into decompositions.
This is an intermediate commit that is probably broken in multiple
ways.

Motivation for the change: We need to support empty decompositions to
deal with emp base rules as well as with non-progress recursive rules
that impose pure constraints.

Instead of a hacky solution with a second type of EntailmentContext, I
decided to move usage info and pure constraints to the decompositions
to then be able to explicitly allow an empty sets of contexts within
the decomposition.

This change also makes sense from a conceptual point of view, as pure
constraints and usage of all (non-placeholder) variables have to be
consistent across all contexts in a decomposition anyway.

This leads to significant algorithmic changes, however. For example,
placeholders are now unique at the level of the decomposition, not the
context, which also complicates the computation of the placeholder
normalform; and composition of contexts and decompositions had to be
rewritten almost in its entirety.

Work in progress (subject to change):
The new context composition is more permissive than the original one:
Unless this leads to double allocation, we allow merging of contexts
even if the root variables of the matching calls don't
match. Whenever (non-placeholder) parameters involved in the merge
don't match, we add this fact as "missing" equality to the pure
constraints.
This way, if the call context provides the missing equality, we can
use these additional parse trees to construct the final state. I
suspect that this kind of approach is necessary for soundness of the
full fragment with pure constraints. If this turns out not to be the
case, I might switch back to a more restrictive (and thus more
efficient) composition operation.