Decomposition composition: Propagate unification into later composition steps.
Since the var usage info is now per decomposition, not per context, we
have to ensure that all contexts involved in the merge process agree
on the var usage info. In particular, this holds for contexts that
have not been processed. Therefore we now update the usage info of
unprocessed processes after each composition step.