Fix: In restrictTo, only throw out non-null vars that don't occur.
This is consistent with treating null as an implicit free vars. (Free
vars are never dropped either, because they necessarily occur in the
profile.)

Motivation behind the fix: Otherwise the behavior in decomposition
composition is not consistent with the behavior in local profiles,
yielding profiles that are identical apart from the null constraints.