Weaker decomposition invariants: Just make sure we drop placeholders.
Previously, we required that the pure constraints cannot contain any
variables that are not used in the contexts. Consequently, we always
dropped pure constraints of vars not in the contexts.
However, that's only sound (and required) for placeholders; for names,
it's actually important to keep all constraints around.  Simplest
example: Decomposition for emp : {x_1 = x_2}.  We need to keep the
constraint x_1 = x_2, even though these variables obviously don't
occur in the (zero) contexts that make up the decomposition.