Replace all occurrences of SpeculativeUpdate with PureAtomUpdate.
That's because it's always fine for equalities with placeholders to
occur, as long as we don't add them to the speculated equalities --
which is exactly the semantics of the PureAtomUpdate.

In the next commit, I'll rename it to the now unused name
SpeculativeUpdate, as this better fits the functionality of the class.