New class for profiles of non-decomposable models.
This class tracks the ensured constraints of the underlying models,
thus ensuring that we only consider this profile for top-level
entailment checking if the top-level constraints do not contradict the
constraints of the profile.

See examples/entailment/empty_lists/ls0-combination.hrs for an example
where this kind of tracking is necessary.

Current problem: The way we add constraints to the sink state (in
profile composition) is not sound, since we have no way to recover the
ensured constraints from constraints that contain speculation.

Possible solution: Also track shared, ensured constraints in
decomposable profiles. (Having shared constraints in both profile
classes would also allow us to merge the two profile classes back into
one if we wanted to, but I think it's cleaner/clearer to keep the
distinction.)