Profiles now have a set rather than a seq of params.
This works fine because we only ever rename profiles that have the
default FVs as params, so there's no need to store the order.

This change will make it easier to compute partial composition
results (combining just a few rather than all call profiles, for
example) -- we can just take the union of the params of the involved
profiles rather than having to assign an arbitrary order.